gcc warning-ok kiszedese

+++ src/autostart-prg.c
@@ -76,7 +76,7 @@ static autostart_prg_t * load_prg(const char *file_name, fileio_info_t *finfo, l
     /* check range */
     end = prg->start_addr + prg->size - 1;
     if (end > 0xffff) {
-        log_error(log, "Invalid size of '%s': %d", file_name, prg->size);
+        log_error(log, "Invalid size of '%s': %d", file_name, (int) prg->size);
         return NULL;
     }
 
@@ -284,7 +284,7 @@ int autostart_prg_perform_injection(log_t log)
     
     log_message(log, "Injecting program data at $%04x (size $%04x)", 
                 prg->start_addr,
-                prg->size);
+                (int) prg->size);
                 
     /* store data in emu memory */
     for (i = 0; i < prg->size; i++) {
